+++ src/commands/ai/hugging_face.ts
