+++ src/commands/ai/sdxl_turbo.ts
